[refactor] jackson deserializer -> core/json 패키지로 이동 - #94
